--- a/src/Compilers/Core/Portable/Diagnostic/Diagnostic.cs
+++ b/src/Compilers/Core/Portable/Diagnostic/Diagnostic.cs
@@ -341,7 +341,10 @@ namespace Microsoft.CodeAnalysis
 
             AttributeData attribute;
             var suppressMessageState = new SuppressMessageAttributeState(compilation);
-            if (!suppressMessageState.IsDiagnosticSuppressed(this, getSemanticModel: tree => compilation.GetSemanticModel(tree), out attribute))
+            if (!suppressMessageState.IsDiagnosticSuppressed(
+                    this,
+                    getSemanticModel: (compilation, tree) => compilation.GetSemanticModel(tree),
+                    out attribute))
             {
                 attribute = null;
             }