--- a/src/Compilers/Core/Portable/DiagnosticAnalyzer/CompilationWithAnalyzers.cs
+++ b/src/Compilers/Core/Portable/DiagnosticAnalyzer/CompilationWithAnalyzers.cs
@@ -1156,7 +1156,7 @@ namespace Microsoft.CodeAnalysis.Diagnostics
             }
 
             var suppressMessageState = new SuppressMessageAttributeState(compilation);
-            var semanticModelsByTree = new ConcurrentDictionary<SyntaxTree, SemanticModel>();
+            var semanticModelsByTree = new Dictionary<SyntaxTree, SemanticModel>();
             foreach (var diagnostic in diagnostics)
             {
                 if (diagnostic != null)
@@ -1169,11 +1169,12 @@ namespace Microsoft.CodeAnalysis.Diagnostics
                 }
             }
 
-            SemanticModel getSemanticModel(SyntaxTree tree)
+            SemanticModel getSemanticModel(Compilation compilation, SyntaxTree tree)
             {
                 if (!semanticModelsByTree.TryGetValue(tree, out var model))
                 {
-                    model = semanticModelsByTree.GetOrAdd(tree, compilation.GetSemanticModel(tree));
+                    model = compilation.GetSemanticModel(tree);
+                    semanticModelsByTree.Add(tree, model);
                 }
 
                 return model;